docs: Format `entrypoint` documentation
Format doc-comments and comment to be 100 characters wide.

Use quotes when referencing feature names.  `` `"std"` feature ``,
rather than `` "`std`" feature ``, or `` `std` feature ``.
This seems to be the common convention.

Also added a few automatic links in the doc comments, where previously
it was just quoted name of a type.